django/requirements.txt
Django==2.0.6
psycopg2==2.7.5
djangorestframework==3.8.2
django-cors-headers==2.2.0
django-model-utils==3.1.2
# git+git://github.com/tiran/defusedxml
SQLAlchemy==1.2.8
# pyparsing==2.1.10
requests==2.19.1
# python-memcached==1.59
social-auth-app-django==2.1.0
social-auth-core==1.7.0
raven==6.9.0
tweepy==3.6.0
celery==4.2.1
django-celery-results==1.0.1